# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Entrophy API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/guilds/{guildId}/economy/config": {
      "get": {
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{17,20}$"
            },
            "in": "path",
            "name": "guildId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      },
      "put": {
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "currencyName": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 32
                  },
                  "currencySymbol": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 8
                  },
                  "dailyMinAmount": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000
                  },
                  "dailyMaxAmount": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000
                  },
                  "streakBonusPerDay": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 10000
                  },
                  "streakBonusMax": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 1000000
                  },
                  "giveMinAmount": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000000000
                  },
                  "giveMaxAmount": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 1000000000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d{17,20}$"
            },
            "in": "path",
            "name": "guildId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ]
}